// src/payload-types.ts
export interface Inquiry {
  id: number;
  name?: string | null;
  mobile?: string | null;
  email?: string | null;
  travellers?: string | null;
  destination?: string | null;
  type?: string | null;
  updatedAt: string;
  createdAt: string;
}